fixing loanReducer and adding window confirms

+++ frontend/src/components/Loan.jsx
@@ -8,10 +8,12 @@ const Loan = ( {book }) => {
     const dispatch = useDispatch()
 
     const returnBook = async (id) => {
+        if (window.confirm(`Return book ${book.title} by ${book.author}?`)) {
         await dispatch(removeLoan(id))
         await dispatch(getUserData()) 
         await dispatch(setNotification( {data: `${book.title} returned`, type: 'info'}, 3000))
-       }
+        }
+    }
 
     const renewCurrentLoan = async (updatedObject) => {
         console.log("renewing")
